third_party/blink/web_tests/external/wpt/screen-orientation/orientation-reading-expected.txt

This is a testharness.js-based test.
PASS Test screen.orientation properties
FAIL Test screen.orientation default values. assert_true: expected true got false
FAIL Test the orientations and associated angles assert_equals: Secondary orientation 1 angle must be 180 expected 180 but got 0
FAIL Test that screen.orientation properties are not writable Cannot set property type of #<ScreenOrientation> which has only a getter
PASS Test that screen.orientation is always the same object
FAIL Test that screen.orientation values change if the orientation changes assert_true: expected event expected true got false
Harness: the test ran to completion.
